[LocationBar] Major cleanup. New code to change color of text parts
The code to change color of host of url is completely new.
It doesn't use any paint hacks anymore. That means text in
locationbar will now be always rendered pixel perfect, as it would
with normal QLineEdit.

Cleanup in LocationCompleter. It is now using QCompleter in inline
mode to complete domain. It is still not ideal, eg. it refreshes
and show the domain completion with delay (job run in separate thread),
and sometimes it misses it completely.

WebSearchBar: Always complete when receiving suggestions. Fix issue
when suggestions were only showed after typing at least 2 characters.