[BookmarksToolbarButton] Final fix for rendering issues with various styles
It now works correctly with Windows, Gtk+, Oxygen and Fusion theme with
both Qt4 and Qt5.
The only theme that doesn't work correctly is Bespin, because it draws
icon size animation upon hover, but we are drawing the icon and text
manually. I think there is not much to do with it actually...