MouseGestures: configurable mouse gesture button
UI changes:
* the Settings dialog offers a combo box to select one of 3
  mouse buttons
* change is permanently saved
* OK button is added

The motivation is allow right mouse button for gestures. This interferes
with the context menu and needed to do custom handling of this event.

Context menu newly pops on mouse right button release unless a gesture
was recognized. In that case the gesture consumes the event and no
context menu appears.

Closes #386