Add application category for OS X
Because /Applications is typically not subfoldered (updaters don't like it when it is) OS X has had a separate view option View --> Arrange By --> Application Category which simplifies viewing apps. Unfortunately if LSApplicationCategoryType is missing from the app's info.plist, the app will be sorted into "Other" at the bottom. public.app-category.productivity is the category for web browsers.