GreaseMonkey: Fix GM_listValues() listing nothing
The function GM_listValues() got broken when the source file "values.js"
was minimized into "values.min.js" because the optimizing tool was not
aware that the string containing the placeholder (%1) is not final and
will be changed later. It optimized away the expression which gets the
length of the string and replaced it with its static value (2). This was
a problem because not the length of the placeholder (%1) but the length
of the current script id is needed to extract prefixes from the keys in
the local storage. Failing to correctly extract the prefixes prevented
finding the entries for the current script. This was fixed by comparing
prefixes without extracting new strings. It works around the need to get
the length of the id string and therefore the breaking optimization.

An unrelated and redundant variable declaration was also removed.