Disable save in menu for about and qupzilla scheme
The pages addressed with the about and qupzilla scheme can currently not
be saved. This means that selecting the save entry in the file menu has
no effect. To prevent users from concluding that the saving should work
but actually does not, the save entry is now disabled for these schemes.