Fix duplicated RSS feed dialog having wrong parent
The warning dialog about duplicated RSS feeds was opened with the main
application window being the parent instead of the RSS feed window. This
opened the warning dialog behind the RSS feed window which was confusing
because the window needed to be moved away to find and close the dialog.