Mac: Resume prior behavior on Makefile and continue to compile on absence
* Qt 5.6 doesn't appear to support any other scopes other than `debug` and `release`
* Nothing I could find in `.../mkspecs` that was applicable
* Predefined targets don't seem to have a `build` or `clean` phase state
* Qt `system` calls don't appear to have access to `MAKECMDGOALS`
* Qt `QMAKE_EXTRA_TARGETS` could possibly work if the environment scope was the same... see prior list item

So basically at this time it is an unsupported feature in current Qt... so get to wait until something else pops up or newer functionality.

It is still a good idea to allow dev testing with openssl under Mac ... otherwise it's saying Linux and Windows can test `master` on openssl but Mac users can't.