VerticalTabs: Reject drop on last row in pinned tabs model
It seems like a weird behavior in Qt because source model receives
drop on last index (in filter model) as drop on last index in source
model instead of filter model and it can't make a right decision.

Source model:
| 1 | 2 | 3 | 4 | 5 | 6 |
Filter model:
| 1 | 2 | 3 | 4 |
                ^
Dropping here propagates to source model as drop on index 7, not 5.

Let's just ignore this case to workaround it.